# internal/ports/docs/openapi.yaml
components:
  schemas:
    ActionRequest:
      properties:
        action:
          example: reboot
          type: string
        dryRun:
          example: true
          type: boolean
        ip:
          example: 10.0.0.1
          type: string
        model:
          example: F1X
          type: string
        params:
          additionalProperties:
            type: string
          type: object
        port:
          example: 60022
          type: integer
      type: object
    ActionResponse:
      properties:
        action:
          type: string
        command:
          type: string
        completedAt:
          type: string
        dryRun:
          type: boolean
        errors:
          items:
            type: string
          type: array
          uniqueItems: false
        ip:
          type: string
        model:
          type: string
        output:
          type: string
        params:
          additionalProperties:
            type: string
          type: object
        port:
          type: integer
        profile:
          type: string
        requestId:
          type: string
        retryable:
          type: boolean
        sshFailed:
          type: boolean
        success:
          type: boolean
        timestamp:
          type: string
      type: object
    CollectRequest:
      properties:
        ip:
          type: string
        model:
          type: string
        port:
          type: integer
      type: object
    CollectResponse:
      properties:
        arp:
          items:
            $ref: '#/components/schemas/cpe.ArpEntry'
          type: array
          uniqueItems: false
        cpeInfo:
          additionalProperties:
            type: string
          type: object
        dhcpLeases:
          items:
            $ref: '#/components/schemas/cpe.DhcpLease'
          type: array
          uniqueItems: false
        errors:
          items:
            type: string
          type: array
          uniqueItems: false
        ethPorts:
          items:
            $ref: '#/components/schemas/cpe.EthPort'
          type: array
          uniqueItems: false
        interfaces:
          items:
            $ref: '#/components/schemas/cpe.Iface'
          type: array
          uniqueItems: false
        ip:
          type: string
        lanHosts:
          items:
            $ref: '#/components/schemas/cpe.LanHost'
          type: array
          uniqueItems: false
        loadAvg:
          $ref: '#/components/schemas/cpe.ProcLoadAvg'
        model:
          type: string
        port:
          type: integer
        portMap:
          items:
            $ref: '#/components/schemas/cpe.PortMapEntry'
          type: array
          uniqueItems: false
        raw:
          additionalProperties:
            type: string
          type: object
        sfp:
          $ref: '#/components/schemas/cpe.SfpInfo'
        sshFailed:
          type: boolean
        timestamp:
          type: string
        uptime:
          $ref: '#/components/schemas/cpe.UptimeInfo'
        wifi24Assoc:
          items:
            $ref: '#/components/schemas/cpe.WifiAssoc'
          type: array
          uniqueItems: false
        wifi50Assoc:
          items:
            $ref: '#/components/schemas/cpe.WifiAssoc'
          type: array
          uniqueItems: false
        wlanConfig:
          items:
            $ref: '#/components/schemas/cpe.WlanRow'
          type: array
          uniqueItems: false
      type: object
    ErrorBody:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        message:
          type: string
        requestId:
          type: string
        retryable:
          type: boolean
        status:
          type: integer
      type: object
    ErrorEnvelope:
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      type: object
    HealthStatus:
      properties:
        status:
          type: string
      type: object
    SwitchPortMAC:
      properties:
        mac:
          type: string
        vlan:
          type: integer
      type: object
    SwitchPortOptics:
      properties:
        currentMa:
          type: number
        rxPowerDbm:
          type: number
        temperatureC:
          type: number
        txPowerDbm:
          type: number
        voltageV:
          type: number
      type: object
    SwitchPortRequest:
      properties:
        host:
          example: 10.160.25.72
          type: string
        includeMacs:
          example: true
          type: boolean
        port:
          example: 6/2
          type: string
        portGroup:
          example: TAFAALLERSTADAR2S003P20
          type: string
        reverseDirection:
          example: false
          type: boolean
      type: object
    SwitchPortResponse:
      properties:
        errors:
          items:
            type: string
          type: array
          uniqueItems: false
        host:
          type: string
        pollTime:
          type: string
        ports:
          items:
            $ref: '#/components/schemas/SwitchPortStatus'
          type: array
          uniqueItems: false
        retryable:
          type: boolean
        snmpFailed:
          type: boolean
        snmpPort:
          type: integer
        switch:
          type: string
      type: object
    SwitchPortStatus:
      properties:
        adminStatus:
          type: string
        description:
          type: string
        dsBytes:
          type: integer
        dsDataMB:
          type: number
        dsPacketErrors:
          type: integer
        duplex:
          type: string
        ifIndex:
          type: integer
        macs:
          items:
            $ref: '#/components/schemas/SwitchPortMAC'
          type: array
          uniqueItems: false
        operStatus:
          type: string
        optics:
          $ref: '#/components/schemas/SwitchPortOptics'
        pollTime:
          type: string
        port:
          type: string
        speedMbps:
          type: integer
        statusDuration:
          type: string
        statusSeconds:
          type: integer
        switch:
          type: string
        usBytes:
          type: integer
        usDataMB:
          type: number
        usPacketErrors:
          type: integer
      type: object
    cpe.ArpEntry:
      properties:
        host:
          type: string
        iface:
          type: string
        ipv4:
          type: string
        mac:
          type: string
        state:
          type: string
      type: object
    cpe.DhcpLease:
      properties:
        expirySeconds:
          type: integer
        hostname:
          type: string
        ipv4:
          type: string
        mac:
          type: string
      type: object
    cpe.EthPort:
      properties:
        duplex:
          type: string
        enable:
          type: integer
        interface:
          type: string
        raw:
          type: string
        speed:
          type: integer
        status:
          type: string
      type: object
    cpe.Iface:
      properties:
        broadcast:
          type: string
        hwaddr:
          type: string
        ipv4:
          type: string
        ipv6LinkLocal:
          type: string
        mtu:
          type: integer
        name:
          type: string
        netmask:
          type: string
        rxBytes:
          type: integer
        rxDropped:
          type: integer
        rxErrors:
          type: integer
        rxPackets:
          type: integer
        txBytes:
          type: integer
        txDropped:
          type: integer
        txErrors:
          type: integer
        txPackets:
          type: integer
      type: object
    cpe.LanHost:
      properties:
        addressSource:
          type: string
        connectionType:
          type: string
        ipv4:
          type: string
        ipv6:
          type: string
        mac:
          type: string
        name:
          type: string
      type: object
    cpe.PortMapEntry:
      properties:
        macNoColons:
          type: string
        portCodeHex:
          type: string
      type: object
    cpe.ProcLoadAvg:
      properties:
        lastPid:
          type: integer
        load:
          items:
            type: number
          type: array
          uniqueItems: false
        procs:
          type: string
        raw:
          type: string
      type: object
    cpe.SfpInfo:
      properties:
        duplex:
          type: string
        linkState:
          type: string
        partNumber:
          type: string
        present:
          type: boolean
        raw:
          type: string
        rxPowerDbm:
          type: number
        serialNumber:
          type: string
        speedMbit:
          type: integer
        temperatureC:
          type: number
        txPowerDbm:
          type: number
        vendorName:
          type: string
      type: object
    cpe.UptimeInfo:
      properties:
        clockTime:
          type: string
        loadAvg:
          items:
            type: number
          type: array
          uniqueItems: false
        raw:
          type: string
        upText:
          type: string
      type: object
    cpe.WifiAssoc:
      properties:
        address:
          type: string
        level:
          type: integer
        rateKbps:
          type: integer
        rssi:
          type: integer
        snr:
          type: integer
      type: object
    cpe.WlanRow:
      properties:
        band:
          type: string
        bandwidth:
          type: string
        channel:
          type: string
        enable:
          type: integer
        index:
          type: integer
        maxDevices:
          type: integer
        pmf:
          type: string
        pskValue:
          type: string
        raw:
          type: string
        securityMode:
          type: string
        ssid:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      description: Paste the API key value directly (a `Bearer ` prefix is also accepted).
      in: header
      name: Authorization
      type: apiKey
externalDocs:
  description: ""
  url: ""
info:
  description: Collect diagnostics from CPEs over SSH and from switches over SNMP.
  title: Device API
  version: "1.0"
openapi: 3.1.0
paths:
  /healthz:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
          description: OK
      summary: Liveness probe
      tags:
      - Health
  /readyz:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
          description: OK
      summary: Readiness probe
      tags:
      - Health
  /v1/cpe/actions:
    post:
      description: Executes a model-aware action (reboot, semi_reset, factory_reset)
        over SSH. Supports dryRun.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/ActionRequest'
                description: Action request
                summary: payload
        description: Action request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Forbidden
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Request Entity Too Large
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Too Many Requests
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
          description: Bad Gateway
      security:
      - BearerAuth: []
      summary: Perform an action on a CPE
      tags:
      - CPE
  /v1/cpe/collect:
    get:
      description: Returns parsed diagnostics collected over SSH.
      parameters:
      - description: Target CPE IP
        in: query
        name: ip
        schema:
          type: string
      - description: Target SSH port (defaults per model)
        in: query
        name: port
        schema:
          type: integer
      - description: Model (VANTIVA/F1X/EWA/FMG/P2812/VMG/AX/EX)
        in: query
        name: model
        schema:
          type: string
      - description: Include raw command output
        in: query
        name: raw
        schema:
          type: boolean
      - description: Include cleartext PSK values
        in: query
        name: includePsk
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Forbidden
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Request Entity Too Large
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Too Many Requests
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectResponse'
          description: Bad Gateway
      security:
      - BearerAuth: []
      summary: Collect CPE diagnostics
      tags:
      - CPE
    post:
      description: Same as GET but takes a JSON body. Returns parsed diagnostics collected
        over SSH.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/CollectRequest'
                description: Target CPE data
                summary: payload
        description: Target CPE data
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Forbidden
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Request Entity Too Large
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Too Many Requests
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectResponse'
          description: Bad Gateway
      security:
      - BearerAuth: []
      summary: Collect CPE diagnostics (JSON body)
      tags:
      - CPE
  /v1/switch/ports:
    get:
      description: Polls a switch (Cisco Catalyst 4500/9400, Huawei S5736) over SNMP
        for interface status, speed, duplex, error counters, byte counters, and transceiver
        optics. Omit port to return all interfaces.
      parameters:
      - description: Switch management IP; ignored when portGroup is set
        in: query
        name: host
        schema:
          type: string
      - description: Interface label, e.g. 6/2 or GigabitEthernet6/2; omit for all
          ports
        in: query
        name: port
        schema:
          type: string
      - description: RADIUS port-group id, e.g. TAFAALLERSTADAR3S003P20; resolves
          host+interface
        in: query
        name: portGroup
        schema:
          type: string
      - description: Include learned MAC addresses (defaults on when a single port
          is requested)
        in: query
        name: macs
        schema:
          type: boolean
      - description: Reverse Ds/Us direction mapping (for core-facing ports)
        in: query
        name: reverse
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwitchPortResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Not Found
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Too Many Requests
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwitchPortResponse'
          description: Bad Gateway
      security:
      - BearerAuth: []
      summary: Poll switch port status and optics over SNMP
      tags:
      - Switch
    post:
      description: Same as GET but takes a JSON body. Omit port to return all interfaces.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/SwitchPortRequest'
                description: Switch poll request
                summary: payload
        description: Switch poll request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwitchPortResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Not Found
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: Too Many Requests
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwitchPortResponse'
          description: Bad Gateway
      security:
      - BearerAuth: []
      summary: Poll switch port status and optics over SNMP (JSON body)
      tags:
      - Switch
servers:
- url: /